+++ ios/Sources/CapacitorCalendarPlugin/Implementation/CapacitorCalendarNew.swift
@@ -383,6 +383,23 @@ class CapacitorCalendarNew: NSObject, EKEventEditViewDelegate, EKCalendarChooser
         }
     }
 
+    func deleteRemindersById(_ input: DeleteRemindersByIdInput) throws -> DeleteRemindersByIdResult {
+        var result = DeleteRemindersByIdResult()
+        input.getIds().forEach { id in
+            do {
+                try ImplementationHelper.deleteReminder(reminderId: id, eventStore: eventStore)
+                result.deleted(id)
+            } catch {
+                result.failed(id)
+            }
+        }
+        return result
+    }
+
+    func deleteReminder(_ input: DeleteReminderInput) throws {
+        try ImplementationHelper.deleteReminder(reminderId: input.getId(), eventStore: eventStore)
+    }
+
     func calendarChooserDidFinish(_ calendarChooser: EKCalendarChooser) {
         var selectCalendarsWithPromptCancellable: AnyCancellable?
         selectCalendarsWithPromptCancellable = self.selectCalendarsWithPromptResultEmitter.sink { promise in