+++ src/schemas/interfaces/delete-reminder-options.ts
@@ -1,0 +1,9 @@
+/**
+ * @since 7.1.0
+ */
+export interface DeleteReminderOptions {
+  /**
+   * @since 7.1.0
+   */
+  id: string;
+}